Renaming processor to `add_id`
