+++ libbeat/docs/processors-list.asciidoc
@@ -14,6 +14,9 @@ endif::[]
 ifndef::no_add_host_metadata_processor[]
 * <<add-host-metadata,`add_host_metadata`>>
 endif::[]
+ifndef::no_add_id_processor[]
+* <<add-id,`add_id`>>
+endif::[]
 ifndef::no_add_kubernetes_metadata_processor[]
 * <<add-kubernetes-metadata,`add_kubernetes_metadata`>>
 endif::[]
@@ -92,9 +95,6 @@ endif::[]
 ifndef::no_truncate_fields_processor[]
 * <<truncate-fields, `truncate_fields`>>
 endif::[]
-ifndef::no_uuid_processor[]
-* <<uuid,`uuid`>>
-endif::[]
 //# end::processors-list[]
 
 //# tag::processors-include[]
@@ -109,6 +109,9 @@ include::{libbeat-processors-dir}/actions/docs/add_fields.asciidoc[]
 endif::[]
 ifndef::no_add_host_metadata_processor[]
 include::{libbeat-processors-dir}/add_host_metadata/docs/add_host_metadata.asciidoc[]
+endif::[]
+ifndef::no_add_id[]
+include::{libbeat-processors-dir}/add_id/docs/add_id.asciidoc[]
 endif::[]
 ifndef::no_add_kubernetes_metadata_processor[]
 include::{libbeat-processors-dir}/add_kubernetes_metadata/docs/add_kubernetes_metadata.asciidoc[]
@@ -187,9 +190,6 @@ include::{libbeat-processors-dir}/timestamp/docs/timestamp.asciidoc[]
 endif::[]
 ifndef::no_truncate_fields_processor[]
 include::{libbeat-processors-dir}/actions/docs/truncate_fields.asciidoc[]
-endif::[]
-ifndef::no_uuid_processor[]
-include::{libbeat-processors-dir}/uuid/docs/uuid.asciidoc[]
 endif::[]
 
 //# end::processors-include[]
