+++ libbeat/processors/add_id/docs/add_id.asciidoc
@@ -1,12 +1,12 @@
-[[uuid]]
-=== Generate UUID for an event
+[[add_id]]
+=== Generate an ID for an event
 
-The `uuid` processor generates a random but roughly ordered UUID for an event.
+The `add_id` processor generates a unique ID for an event.
 
 [source,yaml]
 -----------------------------------------------------
 processors:
- - elasticsearch_id: ~
+ - add_id: ~
 -----------------------------------------------------
 
 The following settings are supported:
