+++ libbeat/processors/add_id/generator/errors.go
