+++ libbeat/processors/add_id/generator/es_generator.go
