+++ libbeat/processors/add_id/generator/es_generator_test.go
