+++ libbeat/processors/add_id/generator/generator_test.go
