+++ libbeat/processors/add_id/generator/mac_test.go
